\name{data}
\alias{chol_RNA}
\docType{data}
\title{
The Cholangiocarcinoma mRNA expression profile
}
\description{
The Cholangiocarcinoma mRNA expression profile data from TCGA
}
\usage{data("data")}
\format{
  The format is:
function (..., list = character(), package = NULL, lib.loc = NULL,
    verbose = getOption("verbose"), envir = .GlobalEnv, overwrite = TRUE)
}

\examples{
data(data)
## maybe str(data) ; plot(data) ...
}
\keyword{datasets}
